add a dev mode to run the front-end, fixed bugs

+++ project-inventory-api/src/main/java/com/project_inventory/project_inventory_api/Controller/InventoryController.java
@@ -96,7 +96,8 @@ public class InventoryController {
                 + inventoryTransferDTO.getSourceType() + " to " + inventoryTransferDTO.getDestinationType());
         InventoryValues sourceType = getInventoryType(inventoryTransferDTO.getSourceType());
         InventoryValues destinationType = getInventoryType(inventoryTransferDTO.getDestinationType());
-        inventoryService.TransferInventoryItem(sourceType, destinationType, inventoryTransferDTO.getInventoryId());
+        inventoryService.TransferInventoryItem(sourceType, destinationType,
+                inventoryTransferDTO.getInventoryId().longValue());
         return "Transferred inventory item";
     }
