+++ project-inventory-api/src/main/java/com/project_inventory/project_inventory_api/Model/InventoryTransferDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class InventoryTransferDTO {
-    private Long inventoryId;
+    private Integer inventoryId;
     private String sourceType;
     private String destinationType;
 }
